refactor: 식당, 마켓, 자유게시판 통일화/ 중복 로직 제거/ 중복 컴포넌트 제거/ 링크 재설정
